feat: userDocsButton
